backend/requirements.txt
fastapi
uvicorn
sqlmodel
pwdlib[argon2]
PyJWT
python-multipart
python-dotenv

